Rename Utils files to {Core,LinuxTracing,Service}Utils
Despite the different modules, the homonymy of these files made it sometimes
annoying to deal with them.
