+++ OrbitCore/CoreUtils.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "Utils.h"
+#include "CoreUtils.h"
 
 #ifdef _WIN32
 // clang-format off
