+++ OrbitGl/App.cpp
@@ -21,6 +21,7 @@
 #include "CallStackDataView.h"
 #include "Callstack.h"
 #include "CaptureWindow.h"
+#include "CoreUtils.h"
 #include "Disassembler.h"
 #include "DisassemblyReport.h"
 #include "FunctionsDataView.h"
@@ -44,7 +45,6 @@
 #include "StringManager.h"
 #include "SymbolHelper.h"
 #include "TimerInfosIterator.h"
-#include "Utils.h"
 #include "capture_data.pb.h"
 #include "preset.pb.h"
 #include "symbol.pb.h"
