+++ OrbitGl/Disassembler.h
@@ -6,7 +6,7 @@
 
 #include <string>
 
-#include "Utils.h"
+#include "CoreUtils.h"
 #include "absl/strings/str_format.h"
 
 class Disassembler {
