+++ OrbitLinuxTracing/TracerThread.h
@@ -19,13 +19,13 @@
 
 #include "ContextSwitchManager.h"
 #include "GpuTracepointEventProcessor.h"
+#include "LinuxTracingUtils.h"
 #include "ManualInstrumentationConfig.h"
 #include "PerfEvent.h"
 #include "PerfEventProcessor.h"
 #include "PerfEventReaders.h"
 #include "PerfEventRingBuffer.h"
 #include "TracepointCustom.h"
-#include "Utils.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "capture.pb.h"
