+++ OrbitService/ProcessList.cpp
@@ -12,7 +12,7 @@
 
 #include "ElfUtils/ElfFile.h"
 #include "OrbitBase/Logging.h"
-#include "Utils.h"
+#include "ServiceUtils.h"
 
 namespace orbit_service {
